refactor: utils 패키지 global 패키지 안으로 이동
