+++ src/main/java/org/programmers/springorder/controller/WalletController.java
@@ -4,8 +4,8 @@ import org.programmers.springorder.console.Console;
 import org.programmers.springorder.constant.Message;
 import org.programmers.springorder.constant.WalletMenuType;
 import org.programmers.springorder.dto.wallet.WalletRequestDto;
+import org.programmers.springorder.global.handler.ExceptionHandler;
 import org.programmers.springorder.service.WalletService;
-import org.programmers.springorder.utils.ExceptionHandler;
 import org.springframework.stereotype.Controller;
 
 import java.util.UUID;
