+++ src/main/java/org/programmers/springorder/global/handler/ExceptionHandler.java
@@ -1,4 +1,4 @@
-package org.programmers.springorder.utils;
+package org.programmers.springorder.global.handler;
 
 import java.util.function.Consumer;
 import java.util.function.Supplier;
