+++ src/main/java/org/programmers/springorder/repository/voucher/JdbcVoucherRepository.java
@@ -13,8 +13,8 @@ import org.springframework.stereotype.Repository;
 import javax.sql.DataSource;
 import java.util.*;
 
-import static org.programmers.springorder.utils.UUIDUtil.toUUID;
-import static org.programmers.springorder.utils.UUIDUtil.uuidToBytes;
+import static org.programmers.springorder.global.utils.UUIDUtil.toUUID;
+import static org.programmers.springorder.global.utils.UUIDUtil.uuidToBytes;
 
 @Repository
 @Primary
